requirements.txt
torch>=2.0.0
torchaudio>=0.9.0
transformers>=4.35.0  
pyannote.audio>=3.0.0
librosa>=0.10.0
numpy>=1.24.0
tqdm>=4.65.0
pillow>=9.0.0
openai-whisper>=20231117
pydub>=0.25.1
pypdf2>=3.0.1
markdown>=3.4.1
